+++ crates/rust-client/src/store/sqlite_store/sync.rs
@@ -7,7 +7,10 @@ use rusqlite::{params, Connection, Transaction};
 use super::SqliteStore;
 use crate::{
     store::{
-        sqlite_store::{accounts::update_account, notes::apply_note_updates_tx},
+        sqlite_store::{
+            accounts::{lock_account, update_account},
+            notes::apply_note_updates_tx,
+        },
         StoreError,
     },
     sync::{NoteTagRecord, NoteTagSource, StateSyncUpdate},
@@ -98,7 +101,7 @@ impl SqliteStore {
             transactions_to_commit: committed_transactions,
             new_mmr_peaks,
             new_authentication_nodes,
-            updated_onchain_accounts,
+            updated_accounts,
             block_has_relevant_notes,
             transactions_to_discard: discarded_transactions,
             tags_to_remove,
@@ -130,8 +133,12 @@ impl SqliteStore {
         Self::mark_transactions_as_discarded(&tx, &discarded_transactions)?;
 
         // Update onchain accounts on the db that have been updated onchain
-        for account in updated_onchain_accounts {
-            update_account(&tx, &account)?;
+        for account in updated_accounts.updated_onchain_accounts() {
+            update_account(&tx, account)?;
+        }
+
+        for (account_id, _) in updated_accounts.mismatched_offchain_accounts() {
+            lock_account(&tx, *account_id)?;
         }
 
         // Commit the updates