+++ crates/rust-client/src/tests.rs
@@ -112,7 +112,10 @@ async fn insert_basic_account() {
     let fetched_account_data = client.get_account(account.id()).await;
     assert!(fetched_account_data.is_ok());
 
-    let (fetched_account, fetched_account_seed) = fetched_account_data.unwrap();
+    let fetched_account = fetched_account_data.unwrap();
+    let fetched_account_seed = fetched_account.seed().cloned();
+    let fetched_account: Account = fetched_account.into();
+
     // Validate header has matching data
     assert_eq!(account.id(), fetched_account.id());
     assert_eq!(account.nonce(), fetched_account.nonce());
@@ -146,7 +149,10 @@ async fn insert_faucet_account() {
     let fetched_account_data = client.get_account(account.id()).await;
     assert!(fetched_account_data.is_ok());
 
-    let (fetched_account, fetched_account_seed) = fetched_account_data.unwrap();
+    let fetched_account = fetched_account_data.unwrap();
+    let fetched_account_seed = fetched_account.seed().cloned();
+    let fetched_account: Account = fetched_account.into();
+
     // Validate header has matching data
     assert_eq!(account.id(), fetched_account.id());
     assert_eq!(account.nonce(), fetched_account.nonce());
@@ -209,8 +215,8 @@ async fn test_account_code() {
         .insert_account(&account, Some(Word::default()), &AuthSecretKey::RpoFalcon512(key_pair))
         .await
         .unwrap();
-    let (retrieved_acc, _) = client.get_account(account.id()).await.unwrap();
-    assert_eq!(*account.code(), *retrieved_acc.code());
+    let retrieved_acc = client.get_account(account.id()).await.unwrap();
+    assert_eq!(*account.code(), *retrieved_acc.account().code());
 }
 
 #[tokio::test]
@@ -573,5 +579,46 @@ async fn test_import_processing_note_returns_error() {
             .await
             .unwrap_err(),
         ClientError::NoteImportError { .. }
+    ));
+}
+
+#[tokio::test]
+async fn test_no_nonce_change_transaction_request() {
+    let mut client = create_test_client().await.0;
+
+    let account_template = AccountTemplate::BasicWallet {
+        mutable_code: false,
+        storage_mode: AccountStorageMode::Private,
+    };
+
+    client.sync_state().await.unwrap();
+
+    // Insert Account
+    let (regular_account, _seed) = client.new_account(account_template).await.unwrap();
+
+    // Prepare transaction
+
+    let code = "
+        begin
+            push.1 push.2
+            # => [1, 2]
+            add push.3
+            # => [1+2, 3]
+            assert_eq
+        end
+        ";
+
+    let tx_script = client.compile_tx_script(vec![], code).unwrap();
+
+    let transaction_request = TransactionRequest::new().with_custom_script(tx_script).unwrap();
+
+    let transaction_execution_result =
+        client.new_transaction(regular_account.id(), transaction_request).await.unwrap();
+
+    let result = client.testing_apply_transaction(transaction_execution_result).await;
+
+    assert!(matches!(
+        result,
+        Err(ClientError::StoreError(StoreError::AccountHashAlreadyExists(_)))
     ));
 }